use stdlib math/bits
